Reduce documentation to one level
Documenting more than one level of folders is unnecessary for
the top level docs overview. All that does is introduce more
places where manual changes need to be made to keep the tree
in sync with the documented trees (and these two trees were
both already out of sync with the actual project tree). Reducing
it to one level of directory makes this easier to maintain,
as the top level structure of the project should change much
less frequently than the internals.