Restructure the library folder
By default, Ansible adds any library folder that is alongside
the top level site.yml file to the library search path. By
restructuring the library folder to be the proper shape for
Ansible to search, we can remove the need to set the search
path in ansible.cfg or environment variables. This version
of the structure should work stright out of the box.