Updated invgen with prompts, removed superfluous debug Added touch function to create files.
linchpin invgen is rewritten to prompt the user when the options are not provided
with the help of click modules
The debug statements and include statements which are not helpful during the
provisioning
The inventory output was needed to be precreated inorder to render inventories.
Inorder to achieve the same touch function is being written in cli utils.