Initial integration of dummy provider
It seemed a bunch of heavy lifting would be required to set up an
environment capable of doing unit tests which didn't really test the
ansible modules (those tests come later). However, the api and the cli
classes need to be tested as unit tests. The dummy provider gives the
functionality to unit test these packages, modules, classes, and
methods.