Added provide_default filter to provide default values if value is null string
Current built in default jinja filter helps the user to omit/provide default to keys
in the ansible module. However , it doesnt provide a way for user to conditionally
provide a default value to ie., based on the value provided by the output of previous
filter . In order to facilitate the same provide_default filter is being written.