Rework credentials bit for auth_driver
Because res_grp['credentials']['name'] is a two dimensional list, going
two levels deep causes the auth_driver module to fail before setting
the auth_var variable. This is due to the fact that it errors because of
an empty outer list, and the inner array can't even be accessed.

The fix is to set some when vars and default to make it cleaner.