Provide for booting openstack from volume
Allows user to specify a specific volume to boot from, which will remain
intact after instance shut down. This volume must be pre-created and
already be flagged as a bootable volume

Allows user to request a new volume be automatically created on boot,
which will then be auto-terminated when the instance is shut down

Lacks the ability to check required/excluded options from schema. That
checking is handled by the Ansible module