Accommodate ansible 2.4.x warnings with file
Create a localhost file in the workspace if it doesn't exist to
eliminate the ansible warnings. This is a temporary fix.

Note: It also appears that it doesn't matter what the content of the
file contains unless it's an IP address. Putting the word 'localhost' in
the file seems to be the best option for now.