Updated logic to handle destroy logic
The logic of this commit allows users to choose how to provision. By
default, provisioning is now unique for the dummy provider (others
coming soon). However, it's now possible to perform an idempotent
reprovision. This can be done by obtaining the run_id on a target by
target basis.

Within the code, there are two different values that track the current
run and the previous (if requested) run.

  The `run_id` value differs from the `rundb_id`, in that
  the `run_id` is an existing value in the database.
  The `rundb_id` value is created to store the new record.
  If the `run_id` is passed, it is used to collect an existing
  `uhash` value from the given `run_id`, which is in turn used
  to perform an idempotent reprovision, or destroy provisioned
  resources.