Move appropriate 'lp_' methods to linchpin.cli pkg
The following methods belong in the cli package:
- lp_init
- lp_destroy
- lp_up

Leaving the following methods in the linchpin package:
- lp_journal

This creates a clear separation of duties. Most tooling will be calling
the do_action method in the linchpin package. The linchpin.cli package
has a method called _build_and_execute, which reads the PinFile,
topology, layout, and (eventually) hooks. This data is passed as a
dictionary to the do_action method, which calls the correct provisioning
action with the proper Ansible playbook.