Create the dummy.yml playbook
Instead of having one common playbook, linchpin now calls a named
playbook to do the task requested. The first one done in any case is
'dummy'. This dummy playbook is called when the topology declares
'resource_group_type: dummy'.

LinchPin knows to call the linchpin/provision/dummy.yml playbook. Which
in turn calls the common, dummy, gather_resources, and inventory_gen
roles (in that order).

This provides for easy extensibility in the future, where a user could
write a new playbook that could provision a new resource, and add it to
the Ansible path. All that is needed is a new role, and a schema file
(which is already included), and a playbook to call said role.