Fix help output in cli
Wile writing documentation, it was discovered that the help didn't print
when options were being passed, even though an argument wasn't called.
The default tooling was insufficient. To remedy the situation, the
click_default_group file was added, which provides a default command to
call if no arguments are passed to the linchpin cli.