Remove lp_test_workspace, use new workspace
After PR #436 is merged, there will be a new workspace in
docs/source/examples/workspace, which is being used for documentation,
but is also going to have valid data to test against. This PR removes
the reference to the previous workspace and points there.

Additionally, the command to teardown in linchpin is 'destroy'. Fixed
this in the linchpin-test.sh