Load configs from linchpin.constants as defaults
Add new file linchpin.constants, which is the old linchpin.conf, but
everything gets loaded into the context. A linchpin.conf in
the normal locations will continue to override configurations. The
overrides now can be done at the configuration item level, rather than
the section level.

The implementation of the load_config method was removed from the
LinchpinContext class and put into the LinchpinCliContext class. This
was done because of the new _load_constants method, which reads from
linchpin.constants and has an entry for every configuration option.