Enable tx_id in up/destroy
Created a new flag for linchpin.conf, 'use_rundb_for_actions'. Now that
the tx_id is in place, a user can pass either the tx_id (with or
without target(s) for filtering), or the run_id and a single target.

The linchpin journal provides additional functionality for looking up
the transaction (--tx_id 40), or by getting the last few transactions in
the transaction view.