Change version to linchpin v1.5.3a (alpha)
This is a new process wherein near the end of each (3wks currently) sprint,
a linchpin release will be made. Alpha (a), beta (b), or rc#
(release candidate) versions will be pre-release. Anything without a
letter in the version name will be an official release which could be
released at any time during a sprint cycle.