Check that virt-install exists
When provisioning libvirt, the step to "Install VM" is set to
ignore errors.  If virt-install does not exist on the system,
there will not be an error until the VM attempts to get a dhcp
ip address a few steps later. This fix checks that virt-install
exists first so that the error is reported immediately.

Signed-off-by: Ryan Cole <rycole@redhat.com>