Creating new workspaces
instead of having one big workspace, linchpin will now have multiple
workspaces in the examples (with a symlink from the root 'workspaces'
path). In the future, tests, examples, etc. will use workspaces. There
will be a README.rst and a standard set of items to add to each
workspace as well.