Allow removing OpenStack routers with connected networks during destroy
Currently removing an os_network that is connected to a router fails
with "Unable to complete operation on network. There are one or more
ports still in use on the network." This change sets the router interfaces
to an empty list during destroy hence allowing the router and
subsequent os_network delete to complete.