Add minimal tests to libvirt and streamline defaults
Minimal tests for a given provider verify that each role in a provider
will run successfully with only the attributes defined as 'required'
in the schema.  There are currently problems with default values not
being set properly and causing provisioning to fail when optional
attributes are not provided.

This also separates default values for libvirt to a 'defaults.yml'
file which is used to populate res_defs at the beginning of provisioning.
All providers will eventually move to this method in order to minimize
the repeated usage of ansible's 'default()' filter which, while useful,
can lead to the problems described above.

Signed-off-by: 14rcole <rycole@redhat.com>